ncorr_python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ncorr"
version = "2.0.0"
description = "Ncorr 2D - Open-source Digital Image Correlation (DIC) software in Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ncorr Python Translation", email = "ncorr@example.com"}
]
keywords = ["DIC", "digital image correlation", "strain", "displacement", "deformation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pillow>=9.0.0",
    "numba>=0.55.0",
    "opencv-python>=4.5.0",
    "PyQt6>=6.0.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ncorr = "ncorr.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ncorr*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4"]
ignore = ["D100", "D104", "D107"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
